requirements.txt
Flask==3.1.2
chardet==4.0.0
reportlab==4.4.5
pandas==2.3.3
python_docx==1.1.2
xlrd==2.0.1
xlwt==1.3.0
openpyxl==3.1.5
xlsxwriter==3.1.2
python-dotenv==1.2.1
Requests==2.32.5
pillow==12.0.0
groq==0.15.0
telebot==0.0.5
